+++ src/gen/from_str.rs
@@ -173,6 +173,12 @@ impl FromStr for crate::v2::Ignore {
         serde_yaml::from_str(s).map_err(Error::new)
     }
 }
+impl FromStr for crate::v2::UpdateType {
+    type Err = Error;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        serde_yaml::from_str(s).map_err(Error::new)
+    }
+}
 impl FromStr for crate::v2::InsecureExternalCodeExecution {
     type Err = Error;
     fn from_str(s: &str) -> Result<Self, Self::Err> {